LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/waiting.https-expected.txt


PASS waiting is set after installation 
PASS The ServiceWorker objects returned from waiting attribute getter that represent the same service worker are the same objects 
